backend/pyproject.toml
[tool.poetry]
name = "genui-backend"
version = "0.1.0"
description = "GenUI SDK Backend - FastAPI server with WebRTC, voice processing, and AI visualization capabilities"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}, {include = "utils"}, {include = "schemas"}]

[tool.poetry.dependencies]
python = "^3.9"

# Core Framework
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# AI & LLM
openai = "^1.3.0"
anthropic = "^0.8.0"
mcp = "^1.0.0"

# Voice & WebRTC Processing
pipecat-ai = "^0.0.40"
pipecat-ai-small-webrtc = "^0.0.3"

# Computer Vision & Media
opencv-python = "^4.8.0"
numpy = "^1.24.0"

# HTTP & Communication
httpx = "^0.25.0"
websockets = "^12.0"
aiohttp = "^3.9.0"

# Utilities & Environment
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"

# Development Dependencies
[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto" 